% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_bigquery_type}
\alias{default_bigquery_type}
\title{Default BigQuery import/export type}
\usage{
default_bigquery_type()
}
\description{
Returns the default BigQuery import/export type. It defaults to
"direct".
}
\seealso{
\code{\link{bigquery_defaults}}
}
