% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_billing_project_id}
\alias{default_billing_project_id}
\title{Default Google BigQuery Billing Project ID}
\usage{
default_billing_project_id()
}
\description{
Returns the default Google BigQuery billing project ID.
}
\seealso{
\code{\link{bigquery_defaults}}
}
