% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_dataset_location}
\alias{default_dataset_location}
\title{Default Google BigQuery Dataset Location}
\usage{
default_dataset_location()
}
\description{
Returns the default Google BigQuery dataset location. It defaults
to "US".
}
\references{
\url{https://cloud.google.com/bigquery/docs/dataset-locations}
}
\seealso{
\code{\link{bigquery_defaults}}
}
