% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_gcs_bucket}
\alias{default_gcs_bucket}
\title{Default Google BigQuery GCS Bucket}
\usage{
default_gcs_bucket()
}
\description{
Returns the default Google BigQuery GCS bucket.
}
\seealso{
\code{\link{bigquery_defaults}}
}
