% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_materialization_dataset}
\alias{default_materialization_dataset}
\title{Default Google BigQuery Materialization Dataset}
\usage{
default_materialization_dataset()
}
\description{
Returns the default Google BigQuery materialization dataset.
}
\seealso{
\code{\link{bigquery_defaults}}
}
