% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_materialization_project}
\alias{default_materialization_project}
\title{Default Google BigQuery Materialization Project}
\usage{
default_materialization_project()
}
\description{
Returns the default Google BigQuery materialization project.
}
\seealso{
\code{\link{bigquery_defaults}}
}
