% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_project_id}
\alias{default_project_id}
\title{Default Google BigQuery Project ID}
\usage{
default_project_id()
}
\description{
Returns the default Google BigQuery project ID.
}
\seealso{
\code{\link{bigquery_defaults}}
}
