% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_service_account_key_file}
\alias{default_service_account_key_file}
\title{Default Google BigQuery Service Account Key File}
\usage{
default_service_account_key_file()
}
\description{
Returns the default service account key file to use.
}
\references{
\url{https://cloud.google.com/bigquery/docs/authentication/service-account-file}
\url{https://cloud.google.com/docs/authentication/}
\url{https://cloud.google.com/bigquery/docs/authentication/}
}
\seealso{
\code{\link{bigquery_defaults}}
}
