% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{bigquery_defaults}
\alias{bigquery_defaults}
\title{Google BigQuery Default Settings}
\usage{
bigquery_defaults(billingProjectId, datasetLocation)
}
\arguments{
\item{billingProjectId}{Default Google Cloud Platform project ID for billing purposes.
This is the project on whose behalf to perform BigQuery operations.}

\item{datasetLocation}{Default Google BigQuery dataset location ("EU" or "US").}
}
\value{
A \code{list} of set options with previous values.
}
\description{
Sets default values for several Google BigQuery related settings.
}
\seealso{
\code{\link{spark_read_bigquery}}, \code{\link{spark_write_bigquery}},
\code{\link{default_billing_project_id}}, \code{\link{default_dataset_location}}
}
\keyword{connection}
\keyword{database,}
